SOLR-6387: Add better error messages throughout Solr and supply a work around for Java bug #8047340 to SystemInfoHandler: On Turkish default locale, some JVMs fail to fork on MacOSX, BSD, AIX, and Solaris platforms.
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1618672 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/contrib/map-reduce/src/test/org/apache/solr/hadoop/MRUnitBase.java
@@ -18,6 +18,7 @@ package org.apache.solr.hadoop;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Locale;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.conf.Configuration;
@@ -35,13 +36,15 @@ public abstract class MRUnitBase extends SolrTestCaseJ4 {
 
   @BeforeClass
   public static void setupClass() throws Exception {
+    assumeFalse("This test fails on UNIX with Turkish default locale (https://issues.apache.org/jira/browse/SOLR-6387)",
+        new Locale("tr").getLanguage().equals(Locale.getDefault().getLanguage()));
     solrHomeZip = SolrOutputFormat.createSolrHomeZip(new File(RESOURCES_DIR + "/solr/mrunit"));
     assertNotNull(solrHomeZip);
   }
 
   @AfterClass
   public static void teardownClass() throws Exception {
-    solrHomeZip.delete();
+    if (solrHomeZip != null) solrHomeZip.delete();
   }
   
   protected void setupHadoopConfig(Configuration config) throws IOException {
